+++ NEWS.md
@@ -1,4 +1,6 @@
 # keeper 0.1.0 (dev version)
 
+* `depot_export` and `depot_import`: store and retrieve a depot in a file on/from disk.
+* `keep_file` to add a file from disk to a depot.
 * `keep` and `pick` to add and retrieve objects from depot.
 * add basic depot handlers `depots_*` and `depot_*`.